refac: memoized token helper + include base APR for lending
